update for new requirements from Frank
mainly, we want to reproduce the numbers from the gracc home dashboard.

Why are they slightly higher here than on the dashboard?  And why does
the all-science-except-lhc sum not add up quite right?  I really don't
know at this point, but i guess we'll have to look into it.

also, we are dropping the last updated date from the webpage and just
including a unix timestamp in the json, per BB.